feat:fastapi,vLLM 연동문제 해결,상대경로수정,쓰이지않는 코드정리
