--- ai_server/client/__init__.py
@@ -1,9 +1,0 @@
-"""
-vLLM 클라이언트 모듈
-FastAPI 서버에서 vLLM 서버와 통신하기 위한 클라이언트
-"""
-
-from .vllm_client import VLLMClient, VLLMAsyncClient
-from .openai_client import OpenAICompatibleClient
-
-__all__ = ["VLLMClient", "VLLMAsyncClient", "OpenAICompatibleClient"]
