--- ai_server/models/__init__.py
@@ -1,5 +1,0 @@
-"""
-AI 모델 구현 모듈
-"""
-
-from ai_server.models.huggingface_model import PostModel
