+++ ai_server/post/post_prompt.py
@@ -15,25 +15,12 @@ class PostPromptGenerator(BaseModel):
     def create_prompt(self) -> PromptTemplate:
         """프롬프트 템플릿 생성"""
 
-        # CLOVAX 형식 프롬프트
+        #프롬프트 
         template = (
-            "### Instruction:\n"
-            f"위의 {self.content}를 {self.emotion}과 {self.post_type}에 어울리게 짧고 자연스럽게 바꿔줘.\n"       
-            f"- 문장 길이는 {self.content}와 비슷하게 유지해줘. (너무 길거나 짧지 않게)\n"
-            "- 해시태그, 광고 문구, 의미 없는 반복 표현은 제거해줘.\n"
-
-            f"### Input:\n"
-            f"너는 {self.post_type} 동물이야. 그리고 현재 감정은 {self.emotion}와 같아."
-            f"지금부터 {self.content}의 글을 말투 변환할꺼야."
-
-            "### Output:\n"
+            f"### Instruction:\n다음 문장을 {self.post_type}의 {self.emotion}한 말투로 바꿔줘.\n"
+            f"### Input:\n문장: {self.content}\n감정: {self.emotion}\n유형: {self.post_type}\n\n"
+            "### Output:"
         )
-
-        # template = (
-        #     f"### Instruction:\n다음 문장을 {self.post_type}의 {self.emotion}한 말투로 바꿔줘.\n"
-        #     f"### Input:\n문장: {self.content}\n감정: {self.emotion}\n유형: {self.post_type}\n\n"
-        #     "### Output:"
-        # )
 
         return PromptTemplate(
             input_variables=["content", "emotion", "post_type"],
