+++ ai_server/vLLM_client/__init__.py
@@ -1,0 +1,7 @@
+"""
+vLLM 클라이언트 모듈
+"""
+
+from .vllm_client import VLLMAsyncClient, CompletionRequest
+
+__all__ = ["VLLMAsyncClient", "CompletionRequest"]
